generate provider form with the dynmic components
Added a FileComponent and removed the need for custom views
